@@ -24,16 +24,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.2.14",
-    "@angular/common": "^7.2.16",
-    "@angular/core": "^7.2.16",
-    "@angular/forms": "^7.2.16",
+    "@angular/animations": "^11.0.5",
+    "@angular/common": "^11.0.5",
+    "@angular/core": "^11.0.5",
+    "@angular/forms": "^11.0.5",
     "@angular/http": "^7.2.16",
-    "@angular/platform-browser": "^7.2.16",
-    "@angular/platform-browser-dynamic": "^7.2.16",
+    "@angular/platform-browser": "^11.0.5",
+    "@angular/platform-browser-dynamic": "^11.0.5",
     "@angular/pwa": "^0.12.4",
-    "@angular/router": "^7.2.16",
-    "@angular/service-worker": "^7.2.16",
+    "@angular/router": "^11.0.5",
+    "@angular/service-worker": "^11.0.5",
     "@ionic-native/contacts": "^5.22.0",
     "@ionic-native/core": "^5.22.0",
     "@ionic-native/facebook": "^5.22.0",
@@ -48,11 +48,12 @@
     "@ionic-native/status-bar": "^5.22.0",
     "@ionic-native/toast": "^5.22.0",
     "@ionic-native/youtube-video-player": "^5.22.0",
-    "@ionic/angular": "^4.11.10",
+    "@ionic/angular": "^5.5.2",
     "@ionic/angular-server": "^5.4.0-dev.202009102321.26d3971",
     "@ng-toolkit/serverless": "^8.1.0",
     "angular-calendar": "^0.27.21",
     "angularx-qrcode": "^1.6.4",
+    "angularx-social-login": "^3.5.4",
     "aws-serverless-express": "^3.3.6",
     "cordova-android": "^8.1.0",
     "cordova-plugin-accountkit": "1.4.0",
@@ -75,26 +76,27 @@
     "firebase": "^6.6.2",
     "intro.js": "^2.9.3",
     "lodash": "^4.17.15",
-    "ng-lottie": "^0.3.2",
+    "lottie-web": "^5.7.5",
     "ngx-filter-pipe": "^2.1.2",
     "ngx-google-places-autocomplete": "^2.0.4",
+    "ngx-lottie": "^6.4.0",
     "ngx-swiper-wrapper": "^8.0.2",
     "onesignal-cordova-plugin": "^2.8.4",
     "rxjs": "^6.5.4",
     "serverless-api-compression": "^1.0.0",
     "tslib": "^1.11.1",
-    "zone.js": "~0.8.29"
+    "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "~0.13.8",
-    "@angular-devkit/build-angular": "~0.13.8",
-    "@angular-devkit/core": "~7.3.8",
-    "@angular-devkit/schematics": "~7.3.8",
-    "@angular/cli": "~7.3.8",
-    "@angular/compiler": "^7.2.16",
-    "@angular/compiler-cli": "^7.2.16",
-    "@angular/language-service": "^7.2.16",
-    "@ionic/angular-toolkit": "~1.5.1",
+    "@angular-devkit/architect": "~0.1100.5",
+    "@angular-devkit/build-angular": "^0.1100.5",
+    "@angular-devkit/core": "~11.0.5",
+    "@angular-devkit/schematics": "~11.0.5",
+    "@angular/cli": "~11.0.5",
+    "@angular/compiler": "^11.0.5",
+    "@angular/compiler-cli": "^11.0.5",
+    "@angular/language-service": "^11.0.5",
+    "@ionic/angular-toolkit": "^3.0.0",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "^2.0.8",
     "@types/node": "^12.12.29",
@@ -106,20 +108,20 @@
     "cordova-plugin-statusbar": "^2.4.3",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
+    "karma": "~5.1.1",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "^2.1.1",
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "^5.4.3",
-    "ts-node": "~8.3.0",
-    "tslint": "~5.17.0",
-    "typescript": "~3.1.6",
-    "ts-loader": "^6.2.1",
-    "webpack-cli": "^3.3.10",
     "opencollective": "^1.0.3",
+    "protractor": "^7.0.0",
     "serverless": "^1.60.0",
-    "serverless-apigw-binary": "^0.4.4"
+    "serverless-apigw-binary": "^0.4.4",
+    "ts-loader": "^6.2.1",
+    "ts-node": "~8.3.0",
+    "tslint": "~6.1.3",
+    "typescript": "~4.0.5",
+    "webpack-cli": "^3.3.10"
   },
   "description": "An Ionic project",
   "cordova": {
