@@ -116,7 +116,7 @@ const routes: Routes = [
 
 @NgModule({
   imports: [
-    RouterModule.forRoot(routes, { preloadingStrategy: PreloadAllModules, scrollPositionRestoration:'enabled' })
+    RouterModule.forRoot(routes, { preloadingStrategy: PreloadAllModules, scrollPositionRestoration: 'enabled', relativeLinkResolution: 'legacy' })
   ],
   exports: [RouterModule]
 })
