@@ -1,5 +1,5 @@
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { CardpreviewComponent } from './cardpreview.component';
 
@@ -7,7 +7,7 @@ describe('CardpreviewComponent', () => {
   let component: CardpreviewComponent;
   let fixture: ComponentFixture<CardpreviewComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ CardpreviewComponent ],
       schemas: [CUSTOM_ELEMENTS_SCHEMA],
