@@ -1,5 +1,5 @@
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { FilterProductsComponent } from './filter-products.component';
 
@@ -7,7 +7,7 @@ describe('FilterProductsComponent', () => {
   let component: FilterProductsComponent;
   let fixture: ComponentFixture<FilterProductsComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ FilterProductsComponent ],
       schemas: [CUSTOM_ELEMENTS_SCHEMA],
