@@ -1,5 +1,5 @@
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { LocationsPage } from './locations.page';
 
@@ -7,7 +7,7 @@ describe('LocationsPage', () => {
   let component: LocationsPage;
   let fixture: ComponentFixture<LocationsPage>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ LocationsPage ],
       schemas: [CUSTOM_ELEMENTS_SCHEMA],
