@@ -1,5 +1,5 @@
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { AvailcakesPage } from './availcakes.page';
 
@@ -7,7 +7,7 @@ describe('AvailcakesPage', () => {
   let component: AvailcakesPage;
   let fixture: ComponentFixture<AvailcakesPage>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ AvailcakesPage ],
       schemas: [CUSTOM_ELEMENTS_SCHEMA],
