@@ -1,5 +1,5 @@
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { RefernearnPage } from './refernearn.page';
 
@@ -7,7 +7,7 @@ describe('RefernearnPage', () => {
   let component: RefernearnPage;
   let fixture: ComponentFixture<RefernearnPage>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ RefernearnPage ],
       schemas: [CUSTOM_ELEMENTS_SCHEMA],
