@@ -1,5 +1,5 @@
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { SavedAddressPage } from './saved-address.page';
 
@@ -7,7 +7,7 @@ describe('SavedAddressPage', () => {
   let component: SavedAddressPage;
   let fixture: ComponentFixture<SavedAddressPage>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ SavedAddressPage ],
       schemas: [CUSTOM_ELEMENTS_SCHEMA],
