@@ -1,5 +1,5 @@
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { SavedaddressesprofilePage } from './savedaddressesprofile.page';
 
@@ -7,7 +7,7 @@ describe('SavedaddressesprofilePage', () => {
   let component: SavedaddressesprofilePage;
   let fixture: ComponentFixture<SavedaddressesprofilePage>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ SavedaddressesprofilePage ],
       schemas: [CUSTOM_ELEMENTS_SCHEMA],
