@@ -1,5 +1,5 @@
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { OffersPage } from './offers.page';
 
@@ -7,7 +7,7 @@ describe('OffersPage', () => {
   let component: OffersPage;
   let fixture: ComponentFixture<OffersPage>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ OffersPage ],
       schemas: [CUSTOM_ELEMENTS_SCHEMA],
