@@ -1,5 +1,5 @@
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { IdeasPage } from './ideas.page';
 
@@ -7,7 +7,7 @@ describe('IdeasPage', () => {
   let component: IdeasPage;
   let fixture: ComponentFixture<IdeasPage>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ IdeasPage ],
       schemas: [CUSTOM_ELEMENTS_SCHEMA],
