@@ -1,5 +1,5 @@
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { NewsfeedPage } from './newsfeed.page';
 
@@ -7,7 +7,7 @@ describe('NewsfeedPage', () => {
   let component: NewsfeedPage;
   let fixture: ComponentFixture<NewsfeedPage>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ NewsfeedPage ],
       schemas: [CUSTOM_ELEMENTS_SCHEMA],
