@@ -1,5 +1,5 @@
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { TiktokPage } from './tiktok.page';
 
@@ -7,7 +7,7 @@ describe('TiktokPage', () => {
   let component: TiktokPage;
   let fixture: ComponentFixture<TiktokPage>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ TiktokPage ],
       schemas: [CUSTOM_ELEMENTS_SCHEMA],
