@@ -1,5 +1,5 @@
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { VideosPage } from './videos.page';
 
@@ -7,7 +7,7 @@ describe('VideosPage', () => {
   let component: VideosPage;
   let fixture: ComponentFixture<VideosPage>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ VideosPage ],
       schemas: [CUSTOM_ELEMENTS_SCHEMA],
