@@ -1,5 +1,5 @@
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { TutorialsPage } from './tutorials.page';
 
@@ -7,7 +7,7 @@ describe('TutorialsPage', () => {
   let component: TutorialsPage;
   let fixture: ComponentFixture<TutorialsPage>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ TutorialsPage ],
       schemas: [CUSTOM_ELEMENTS_SCHEMA],
